fixed h2 and images on mobile, changed iframe to image
removed embedded video because it wasn't being responsive. changed it to a clickable image

h2 header was too big on mobile so now its smaller

some images weren't sizing right ie the mail in the about page now they're fixed

figcaption is now smaller and grayer

fixed some margin and padding spacing for the top of the pages.